feat: 댓글 CRUD에 JWT를 활용한 유저 권한 적용

+++ src/main/java/dgu/edu/dnaapi/service/NoticeCommentsService.java
@@ -2,6 +2,7 @@ package dgu.edu.dnaapi.service;
 
 import dgu.edu.dnaapi.domain.NoticeComments;
 import dgu.edu.dnaapi.domain.Notices;
+import dgu.edu.dnaapi.domain.User;
 import dgu.edu.dnaapi.domain.dto.noticeComments.NoticeCommentsDeleteDto;
 import dgu.edu.dnaapi.domain.dto.noticeComments.NoticeCommentsResponseDto;
 import dgu.edu.dnaapi.domain.dto.noticeComments.NoticeCommentsSaveDto;
@@ -22,46 +23,55 @@ public class NoticeCommentsService {
     private final NoticesCommentsRepository noticesCommentsRepository;
 
     @Transactional
-    public Long save(NoticeCommentsSaveDto requestDto) {
-        NoticeComments comment = requestDto.toEntity();
+    public Long save(User user, NoticeCommentsSaveDto requestDto) {
+        NoticeComments comment = requestDto.toEntity(user);
+        comment.registerAuthor(user);
+
         Long noticeId = requestDto.getNoticeId();
-        Long parentCommentId = requestDto.getParentCommentId();
 
         Notices notice = noticesRepository.findById(noticeId).orElseThrow(
                 () -> new IllegalArgumentException("해당 글이 없습니다. id=" + noticeId));
         comment.registerNotice(notice);
 
-        if(parentCommentId != null) {
+        if(requestDto.getParentCommentId() != null) {
+            Long parentCommentId = requestDto.getParentCommentId();
             NoticeComments parent = noticesCommentsRepository.findById(parentCommentId).orElseThrow(
                     () -> new IllegalArgumentException("해당 댓글이 없습니다. id=" + parentCommentId));
             comment.registerParent(parent);
-            parent.addChild(comment);
         }
 
         return noticesCommentsRepository.save(comment).getCommentId();
     }
 
     @Transactional
-    public Long update(NoticeCommentsUpdateDto requestDto) {
+    public Long update(Long userId, NoticeCommentsUpdateDto requestDto) {
         Long id = requestDto.getCommentId();
-        NoticeComments notices = noticesCommentsRepository.findById(id).orElseThrow(
+        NoticeComments comment = noticesCommentsRepository.findById(id).orElseThrow(
                 () -> new IllegalArgumentException("해당 댓글이 없습니다. id=" + id));
-        notices.update(requestDto.getContent());
+
+        if(!comment.getAuthor().getId().equals(userId)) {
+            throw new IllegalArgumentException("작성자만 댓글을 수정할 수 있습니다.");
+        }
+        comment.update(requestDto.getContent());
         return id;
     }
 
     public NoticeCommentsResponseDto findById(Long id) {
         NoticeComments entity = noticesCommentsRepository.findById(id).orElseThrow(
-                () -> new IllegalArgumentException("해당 게시글이 없습니다. id=" + id));
+                () -> new IllegalArgumentException("해당 댓글이 없습니다. id=" + id));
         return new NoticeCommentsResponseDto(entity);
     }
 
     @Transactional
-    public Long delete(NoticeCommentsDeleteDto requestDto) {
+    public Long delete(Long userId, NoticeCommentsDeleteDto requestDto) {
         // todo 나중에 대댓글 삭제 조건 처리 등 추가 해야함
         Long deleteId = requestDto.getCommentId();
+        NoticeComments comment = noticesCommentsRepository.findById(deleteId).orElseThrow(
+                () -> new IllegalArgumentException("해당 댓글이 없습니다. id=" + deleteId));
+        if(!comment.getAuthor().getId().equals(userId)) {
+            throw new IllegalArgumentException("작성자만 댓글을 삭제할 수 있습니다.");
+        }
         noticesCommentsRepository.deleteById(deleteId);
-
         return deleteId;
     }
